Refactored common stuff for Rope
