+++ Runtime/Scripts/Rope/Pulley.cs
@@ -1,18 +1,19 @@
 using UnityEngine;
-
 using Force;
 
 namespace Rope
 {
     [RequireComponent(typeof(Rigidbody))]
-    public class Pulley : MonoBehaviour
+    public class Pulley : RopeSystemBase
     {
-        [Header("Connected Bodies")]
+        [Header("Body One")]
         public ArticulationBody ConnectedABOne;
         public Rigidbody ConnectedRBOne;
+        MixedBody EndOne;
+        
+        [Header("Body Two")]
         public ArticulationBody ConnectedABTwo;
         public Rigidbody ConnectedRBTwo;
-        MixedBody EndOne;
         MixedBody EndTwo;
 
         ConfigurableJoint distanceJointOne, distanceJointTwo;
@@ -20,7 +21,6 @@ namespace Rope
 
         [Header("Rope Properties")]
         public float RopeLength;
-        public float RopeDiameter = 0.1f;
 
         [Header("Debug")]
         public float sideOneLimit;
@@ -28,92 +28,12 @@ namespace Rope
         public float ropeVelocity;
         
 
-        Rigidbody AddIneffectiveRB(GameObject o)
-        {
-            Rigidbody rb = o.AddComponent<Rigidbody>();
-            rb.useGravity = false;
-            rb.inertiaTensor = Vector3.one * 1e-6f;
-            rb.drag = 0;
-            rb.angularDrag = 0;
-            rb.mass = 0.1f;
-            return rb;
-        }
-
-        ConfigurableJoint AddSphericalJoint(GameObject o)
-        {
-            ConfigurableJoint joint = o.AddComponent<ConfigurableJoint>();
-            joint.xMotion = ConfigurableJointMotion.Locked;
-            joint.yMotion = ConfigurableJointMotion.Locked;
-            joint.zMotion = ConfigurableJointMotion.Locked;
-            joint.anchor = Vector3.zero;
-            joint.autoConfigureConnectedAnchor = false;
-            joint.connectedAnchor = Vector3.zero;
-            return joint;
-        }
-
-        ConfigurableJoint AddDistanceJoint(GameObject o)
-        {
-            ConfigurableJoint joint = o.AddComponent<ConfigurableJoint>();
-            joint.xMotion = ConfigurableJointMotion.Locked;
-            joint.yMotion = ConfigurableJointMotion.Limited;
-            joint.zMotion = ConfigurableJointMotion.Locked;
-            joint.angularXMotion = ConfigurableJointMotion.Locked;
-            joint.angularYMotion = ConfigurableJointMotion.Locked;
-            joint.angularZMotion = ConfigurableJointMotion.Locked;
-            joint.anchor = Vector3.zero;
-            joint.autoConfigureConnectedAnchor = false;
-            joint.connectedAnchor = Vector3.zero;
-            return joint;
-        }
-
-
-        ConfigurableJoint AttachToPulley(MixedBody end)
-        {
-            Rigidbody baseRB = GetComponent<Rigidbody>();
-
-            // Spherical connection to this object
-            var sphericalToBase = new GameObject("SphericalToBase");
-            sphericalToBase.transform.parent = transform.parent;
-            var sphericalToBaseRB = AddIneffectiveRB(sphericalToBase);
-            var sphericalToBaseJoint = AddSphericalJoint(sphericalToBase);
-            
-            // Linear connection to the previous sphere
-            var distanceToSpherical = new GameObject("DistanceToSpherical");
-            distanceToSpherical.transform.parent = transform.parent;
-            var distanceToSphericalRB = AddIneffectiveRB(distanceToSpherical);
-            var distanceToSphericalJoint = AddDistanceJoint(distanceToSpherical);
-            // Spherical connection to the end object
-            var sphericalToEndJoint = AddSphericalJoint(distanceToSpherical);
-            
-            // Base -> Sphere -> Linear+Sphere -> End
-            sphericalToBaseJoint.connectedBody = baseRB;
-            distanceToSphericalJoint.connectedBody = sphericalToBaseRB;
-            end.ConnectToJoint(sphericalToEndJoint);
-
-            // Add a linerenderer to visualize the rope and its tight/slack state
-            var lr = distanceToSpherical.AddComponent<LineRenderer>();
-            lr.positionCount = 2;
-            lr.material = new Material(Shader.Find("Sprites/Default"));
-            lr.startWidth = RopeDiameter;
-
-            return distanceToSphericalJoint;
-        }
-
-        void UpdateJointLimit(ConfigurableJoint joint, float length)
-        {
-            var jointLimit = new SoftJointLimit
-            {
-                limit = length
-            };
-            joint.linearLimit = jointLimit;
-        }
-
         void Start()
         {
             EndOne = new MixedBody(ConnectedABOne, ConnectedRBOne);
             EndTwo = new MixedBody(ConnectedABTwo, ConnectedRBTwo);
-            distanceJointOne = AttachToPulley(EndOne);
-            distanceJointTwo = AttachToPulley(EndTwo);
+            distanceJointOne = AttachBody(EndOne);
+            distanceJointTwo = AttachBody(EndTwo);
             sideOneLR = distanceJointOne.gameObject.GetComponent<LineRenderer>();
             sideTwoLR = distanceJointTwo.gameObject.GetComponent<LineRenderer>();
             sideOneLimit = Vector3.Distance(EndOne.position, transform.position);
@@ -125,7 +45,6 @@ namespace Rope
 
         void FixedUpdate()
         {
-
             float sideOneDistance = Vector3.Distance(EndOne.position, transform.position);
             float sideTwoDistance = Vector3.Distance(EndTwo.position, transform.position);
