+++ Runtime/Scripts/Rope/RopeSystemBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1b611011a6434594cac12b5a80bce92c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
